refactor: ScheduleTable 컴포넌트의 DnD 기능 개선 및 코드 구조화
